refactor: reduce duplication of UI components
