--- lib/game_display_widget.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-class GameDisplayWidget extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    // TODO: implement build
-    return const SizedBox.shrink();
-  }
-}
