+++ lib/model/setting.dart
@@ -35,4 +35,46 @@ class Setting {
 
   Setting clone() => Setting(this.name,
       value: this.value, header: this.header, enabled: this.enabled);
+
+  /// Type-safe getter for boolean values
+  /// Handles string-to-bool conversion automatically
+  bool getBool() {
+    if (value is String) {
+      return value.toLowerCase() == 'true';
+    }
+    return value as bool? ?? false;
+  }
+
+  /// Type-safe getter for integer values
+  /// Handles string-to-int conversion automatically
+  int getInt() {
+    if (value is String) {
+      return int.tryParse(value) ?? 0;
+    }
+    if (value is double) {
+      return value.toInt();
+    }
+    return value as int? ?? 0;
+  }
+
+  /// Type-safe getter for double values
+  /// Handles string-to-double conversion automatically
+  double getDouble() {
+    if (value is String) {
+      return double.tryParse(value) ?? 0.0;
+    }
+    if (value is int) {
+      return value.toDouble();
+    }
+    return (value as num?)?.toDouble() ?? 0.0;
+  }
+
+  /// Type-safe getter for list values
+  /// Returns empty list if value is not a list
+  List<dynamic> getList() {
+    if (value is List) {
+      return value as List<dynamic>;
+    }
+    return [];
+  }
 }
